modifiying js files
