+++ src/App.js
@@ -1,25 +1,62 @@
-import logo from './logo.svg';
+import * as React from 'react';
+import PropTypes from 'prop-types'; // Ensure prop-types is imported
 import './App.css';
 
-function App() {
+const quoteData = [
+  { text: '“The purpose of our lives is to be happy.”', author: 'Dalai Lama' },
+  { text: '“Life is what happens when you’re busy making other plans.”', author: 'John Lennon' },
+  { text: '“Get busy living or get busy dying.”', author: 'Stephen King' },
+  { text: '“You only live once, but if you do it right, once is enough.”', author: 'Mae West' },
+  { text: '“Many of life’s failures are people who did not realize how close they were to success when they gave up.”', author: 'Thomas A. Edison' },
+];
+
+const QuoteBox = ({ quote, handleNewQuote }) => (
+  <div id="quote-box">
+    <p id="text">{quote.text}</p>
+    <h2 id="author">{quote.author}</h2>
+    <div className="actions">
+      <button
+        id="new-quote"
+        className="button"
+        onClick={handleNewQuote}
+        type="button" // Ensure button type is specified
+      >
+        New Quote
+      </button>
+      <a
+        href="https://twitter.com/intent/tweet"
+        id="tweet-quote"
+        rel="noreferrer"
+        target="_blank"
+      >
+        Tweet
+      </a>
+    </div>
+  </div>
+);
+
+QuoteBox.propTypes = {
+  quote: PropTypes.shape({
+    text: PropTypes.string.isRequired,
+    author: PropTypes.string.isRequired,
+  }).isRequired,
+  handleNewQuote: PropTypes.func.isRequired,
+};
+
+const getRandomIndex = () => Math.floor(Math.random() * quoteData.length);
+
+const App = () => {
+  const [quote, setQuote] = React.useState(quoteData[getRandomIndex()]);
+
+  const handleNewQuote = () => {
+    setQuote(quoteData[getRandomIndex()]);
+  };
+
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+    <div className="main">
+      <QuoteBox handleNewQuote={handleNewQuote} quote={quote} />
     </div>
   );
-}
+};
 
 export default App;
